+++ src/pyscal3/attributes.py
