+++ src/pyscal3/data/annotations.yaml
