+++ src/pyscal3/data/structure_data.yaml
