+++ src/pyscal3/formats/lammps.py
