+++ src/pyscal3/formats/mdtraj.py
