+++ src/pyscal3/neighbor.cpp
