+++ src/pyscal3/sh.cpp
