+++ src/pyscal3/solids.cpp
